#26 :gear: Refined Approach to Hiding from Evil :robot:
In the initial iteration, we ultimately solved the "evil hacker bots on the internet" problem by simply switching to a non-standard SSH port in the firewall-router. But then our internet provider cancelled support for our static IP a week ago, so now we've been forced over to the only (reasonably priced) provider that can offer a static IP (then via 4G mobile broadband router). The thing is, this consumer grade cellular router might look good, but capability-wise it's a pile of *****sh*t. It only supports port forwarding from upstream interface to a downstream IP using same port number— not forwarding port X to port Y —so we actually need to change the port `sshd` uses on the server. Here's the necessary Ansible updates to make this happen! :boom: